renamed "component" files to actual names
