+++ src/components/MonthSelector/index.ts
@@ -1,9 +1,9 @@
-import { connect } from 'react-redux';
+import { connect } from "react-redux";
 
-import MonthSelector from './component';
-import { StoreState } from '../../types/store';
-import { Month } from '../../types';
-import { updateSelectedMonths } from '../../store/actions/current';
+import MonthSelector from "./MonthSelector";
+import { StoreState } from "../../types/store";
+import { Month } from "../../types";
+import { updateSelectedMonths } from "../../store/actions/current";
 
 const mapStateToProps = (state: StoreState) => ({
   selectedMonths: state.current.months,
