+++ src/components/TopBar/index.ts
@@ -1,7 +1,7 @@
-import { connect } from 'react-redux';
+import { connect } from "react-redux";
 
-import TopBar from './component';
-import { StoreState } from '../../types/store';
+import TopBar from "./TopBar";
+import { StoreState } from "../../types/store";
 
 const mapStateToProps = (state: StoreState) => ({
   amountLeft: state.amountLeft,
